// src/api/be/schemas/apiResponseTournamentSummaryDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { TournamentSummaryDto } from './tournamentSummaryDto';

export interface ApiResponseTournamentSummaryDto {
  success?: boolean;
  /** @nullable */
  message?: string | null;
  data?: TournamentSummaryDto | null;
}