// src/api/generated/schemas/tournamentHistoryItemDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * TripBite API
 * OpenAPI spec version: 1.0
 */
import type { DestinationCategory } from './destinationCategory';
import type { RegionCode } from './regionCode';

export interface TournamentHistoryItemDto {
  category: DestinationCategory;
  winnerRegion: RegionCode;
  id: string;
  /** @nullable */
  theme: string | null;
  count: number;
  winnerId: string;
  winnerName: string;
  completedAt: string;
}